refactor: reorganize shared layout components
