+++ src/components/header/AuthHeader.tsx
